Main change is in pso.py. 1) __call__() is introduced for the optimiser, and buzzi() is renamed to run() 2) change in the interface: the evaluate function that is passed to the PSO class must return the fitness (as a tuple), but also the maxError (absolute value of the max error, either absolute or relative) of the current evaluation. (previously, the entire array of errors was returned and the pso calculated maxError on its own.)
Another change is in the Evaluate class:
now its evaluate method returns a tupple: cost, maxErr.

The exapmle SKOPT notebook is updated with a closed loop between
PSO instance and evaluate instance.
Note that the PSO requires a tuple fitness (per objective, but tuple
even in case of one objective).
So the Evaluator instance is wrapped, to make this tupple in the
wrapper, rather than hack the evaluator, which remains a
single objective cost-function.


Former-commit-id: 4b910f7a1bb2465c97f4cbf63c77b011a52b3de5