MAJOR CHANGE: setup.py was bad and is now fixed as follows: * src directory changed to skopt * src/skopt is now skopt/core * importing skopt stuff is done by: import skopt import skopt.core import skopt.dftbutils Therefore, all references to import skopt.*module* and import dftbutils.*module* are changed both in source and test
Minor fixes related to file path-handling (always normapt(expanduser())
now).
Minor change is test -- using ~/anaconda3/python works only on Windows
but not on Linux (~/anaconda3/bin/python...) and in general, we
do not expect user to have anaconda installed.  Still a few outsanding
cases like that may exist, but should be easy to fix in the tests.


Former-commit-id: befdef59251523f8a43df1f7bf9b751acaee0cd4
Former-commit-id: ed9eaa22b2f82a88b09f8f8062270cc21b60ba70